pyproject.toml
[project]
name = "bots"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Jonathan Miller", email = "jon@jmill.dev" }
]
dependencies = [
    "openai>=1.3.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.6.0",
    "py-cord>=2.4.1",
    "logos @ git+https://github.com/edenartlab/logos",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.5.0",
    "black>=23.11.0",
    "mypy>=1.7.0",
    "flake8>=6.1.0",
    "pytest>=7.4.3",
    "discord-ext-autoreload>=0.1.2",
]

[tool.hatch.metadata]
allow-direct-references = true
